+++ R/migration.R
@@ -124,7 +124,9 @@ migration <- function(
   datapack = NULL,
   success.arrays = NULL,
   max.interval = 60,
-  minimum.detections = 2,
+  minimum.detections,
+  min.total.detections = 2,
+  min.per.event = 1,
   start.time = NULL,
   stop.time = NULL,
   speed.method = c("last to first", "last to last"),
@@ -150,6 +152,10 @@ migration <- function(
   detections.y.axis = c("auto", "stations", "arrays"))
 {
 
+# check deprecated argument
+  if (!missing(minimum.detections))
+    stop("'minimum.detections' has been deprecated. Please use 'min.total.detections' and 'min.per.event' instead.", call. = FALSE)
+
 # clean up any lost helpers
   deleteHelpers()
   if (file.exists(paste0(tempdir(), "/actel_debug_file.txt")))
@@ -173,7 +179,8 @@ migration <- function(
 
   aux <- checkArguments(dp = datapack,
                         tz = tz,
-                        minimum.detections = minimum.detections,
+                        min.total.detections = min.total.detections,
+                        min.per.event = min.per.event,
                         max.interval = max.interval,
                         speed.method = speed.method,
                         speed.warning = speed.warning,
@@ -195,6 +202,7 @@ migration <- function(
                         section.order = section.order,
                         detections.y.axis = detections.y.axis)
 
+  min.per.event <- aux$min.per.event
   speed.method <- aux$speed.method
   speed.warning <- aux$speed.warning
   speed.error <- aux$speed.error
@@ -212,7 +220,8 @@ migration <- function(
     ", datapack = ", ifelse(is.null(datapack), "NULL", deparse(substitute(datapack))),
     ", success.arrays = ", ifelse(is.null(success.arrays), "NULL", paste0("c('", paste(success.arrays, collapse = "', '"), "')")),
     ", max.interval = ", max.interval,
-    ", minimum.detections = ", minimum.detections,
+    ", min.total.detections = ", min.total.detections,
+    ", min.per.event = ", paste0("c('", min.per.event, "')"),
     ", start.time = ", ifelse(is.null(start.time), "NULL", paste0("'", start.time, "'")),
     ", stop.time = ", ifelse(is.null(stop.time), "NULL", paste0("'", stop.time, "'")),
     ", speed.method = '", speed.method, "'",
@@ -397,7 +406,8 @@ migration <- function(
     appendTo("debug", paste0("debug: Checking movement quality for tag ", tag,"."))
 
     if (is.na(match(tag, override))) {
-      output <- checkMinimumN(movements = movements[[tag]], tag = tag, minimum.detections = minimum.detections, n = counter)
+      output <- checkMinimumN(movements = movements[[tag]], tag = tag, min.total.detections = min.total.detections,
+                               min.per.event = min.per.event[1], n = counter)
 
       output <- checkUpstream(movements = output, tag = tag, detections = detections.list[[tag]], spatial = spatial,
                               bio = bio, arrays = arrays, GUI = GUI, save.tables.locally = save.tables.locally, n = counter)
@@ -442,7 +452,11 @@ migration <- function(
     appendTo("debug", paste0("debug: Compiling section movements for tag ", tag,"."))
 
     aux <- sectionMovements(movements = movements[[i]], spatial = spatial, valid.dist = attributes(dist.mat)$valid)
-    if (!is.null(aux)) {
+
+    if (!is.null(aux)) { # interesting... why do I have this here but not on residency? hm...
+      aux <- checkMinimumN(movements = aux, tag = tag, min.total.detections = 0, # don't run the minimum total detections check here.
+                           min.per.event = min.per.event[2], n = counter)
+
       output <- checkLinearity(secmoves = aux, tag = tag, spatial = spatial, arrays = arrays, 
                                GUI = GUI, save.tables.locally = save.tables.locally, n = counter)
       return(output)